WIP: lazy load dlls
Lazy load dlls  https://github.com/Azure/azure-webjobs-sdk-script/issues/1637

This discovers the binding types at runtime and just loads those bindings.
Even though the assembly has a static reference to a dll, that dll is not loaded until its used.
So this removed unused dlls from the runtime working set.

It would be good to find a way to remove the static reference.